megano/requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Django==4.2.2
django-filter==22.1
django-js-asset==2.0.0
django-mptt==0.14.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
idna==3.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
Pillow==9.3.0
platformdirs==3.8.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
transliterate==1.10.2
typing_extensions==4.6.3
urllib3==2.0.3
